Modules
Добавил везде импорты и экспорты.
Перетащил переменные в нужные места
